requirements.txt
aniso8601==9.0.1
azure-cognitiveservices-speech==1.18.0
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
ffmpeg-python==0.2.0
Flask==2.0.1
Flask-RESTful==0.3.9
future==0.18.2
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
modules==1.0.0
mysql-connector-python==8.0.26
Naked==0.1.31
praw==7.4.0
prawcore==2.3.0
protobuf==3.17.3
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
six==1.16.0
update-checker==0.18.0
urllib3==1.26.6
Webshot==0.6
websocket-client==1.1.1
Werkzeug==2.0.1
